# frontend/src/lib/api/openapi.json
{
  "components": {
    "schemas": {
...
      "AvailabilityDaysRead": {
        "properties": {
          "asset_id": {
            "format": "uuid",
            "title": "Asset Id",
            "type": "string"
          },
          "days": {
            "items": {
              "$ref": "#/components/schemas/AvailabilityDayRead"
            },
            "title": "Days",
            "type": "array"
          },
          "ends_at": {
            "format": "date-time",
            "title": "Ends At",
            "type": "string"
          },
          "location_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location Id"
          },
          "quantity": {
            "title": "Quantity",
            "type": "integer"
          },
          "starts_at": {
            "format": "date-time",
            "title": "Starts At",
            "type": "string"
          }
        },
        "required": ["asset_id", "location_id", "quantity", "starts_at", "ends_at", "days"],
        "title": "AvailabilityDaysRead",
        "type": "object"
      },
      "AvailabilityHeatmapRead": {
        "properties": {
          "bucket": {
            "title": "Bucket",
            "type": "string"
          },
          "ends_at": {
            "format": "date-time",
            "title": "Ends At",
            "type": "string"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/HeatmapItemRead"
            },
            "title": "Items",
            "type": "array"
          },
          "location_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location Id"
          },
          "starts_at": {
            "format": "date-time",
            "title": "Starts At",
            "type": "string"
          }
        },
        "required": ["starts_at", "ends_at", "bucket", "location_id", "items"],
        "title": "AvailabilityHeatmapRead",
        "type": "object"
      },
...
      "AvailabilityRead": {
        "properties": {
          "available": {
            "title": "Available",
            "type": "boolean"
          },
          "lines": {
            "items": {
              "$ref": "#/components/schemas/AvailabilityLineRead"
            },
            "title": "Lines",
            "type": "array"
          }
        },
        "required": ["available", "lines"],
        "title": "AvailabilityRead",
        "type": "object"
      },
...
      "BookingCreate": {
        "additionalProperties": false,
        "properties": {
          "ends_at": {
            "format": "date-time",
            "title": "Ends At",
            "type": "string"
          },
          "lines": {
            "items": {
              "$ref": "#/components/schemas/BookingLineCreate"
            },
            "minItems": 1,
            "title": "Lines",
            "type": "array"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "person_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Person Id"
          },
          "starts_at": {
            "format": "date-time",
            "title": "Starts At",
            "type": "string"
          },
          "title": {
            "maxLength": 180,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": ["title", "starts_at", "ends_at", "lines"],
        "title": "BookingCreate",
        "type": "object"
      },
...
      "BookingRead": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "ends_at": {
            "format": "date-time",
            "title": "Ends At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "lines": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/BookingLineRead"
            },
            "title": "Lines",
            "type": "array"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "person_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Person Id"
          },
          "requested_by_user_id": {
            "format": "uuid",
            "title": "Requested By User Id",
            "type": "string"
          },
          "starts_at": {
            "format": "date-time",
            "title": "Starts At",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/BookingStatus"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "requested_by_user_id",
          "person_id",
          "title",
          "status",
          "starts_at",
          "ends_at",
          "created_at",
          "notes"
        ],
        "title": "BookingRead",
        "type": "object"
      },
      "BookingStatus": {
        "enum": ["reserved", "cancelled", "checked_out", "completed"],
        "title": "BookingStatus",
        "type": "string"
      },
      "BookingSummaryRead": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "ends_at": {
            "format": "date-time",
            "title": "Ends At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "person_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Person Id"
          },
          "requested_by_user_id": {
            "format": "uuid",
            "title": "Requested By User Id",
            "type": "string"
          },
          "starts_at": {
            "format": "date-time",
            "title": "Starts At",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/BookingStatus"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "requested_by_user_id",
          "person_id",
          "title",
          "status",
          "starts_at",
          "ends_at",
          "created_at",
          "notes"
        ],
        "title": "BookingSummaryRead",
        "type": "object"
      },
      "BookingUpdate": {
        "additionalProperties": false,
        "properties": {
          "ends_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ends At"
          },
          "person_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Person Id"
          },
          "starts_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Starts At"
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BookingStatus"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "title": "BookingUpdate",
        "type": "object"
      },
...
      "CheckoutCreate": {
        "additionalProperties": false,
        "properties": {
          "booking_id": {
            "format": "uuid",
            "title": "Booking Id",
            "type": "string"
          },
          "checked_out_to_user_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Checked Out To User Id"
          },
          "condition_out": {
            "$ref": "#/components/schemas/AssetCondition",
            "default": "unknown"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "required": ["booking_id"],
        "title": "CheckoutCreate",
        "type": "object"
      },
...
      "CheckoutStatus": {
        "enum": ["checked_out", "partially_returned", "returned"],
        "title": "CheckoutStatus",
        "type": "string"
      },
      "CheckoutSummaryRead": {
        "properties": {
          "booking_id": {
            "format": "uuid",
            "title": "Booking Id",
            "type": "string"
          },
          "checked_out_by_user_id": {
            "format": "uuid",
            "title": "Checked Out By User Id",
            "type": "string"
          },
          "checked_out_to_user_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Checked Out To User Id"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "status": {
            "$ref": "#/components/schemas/CheckoutStatus"
          }
        },
        "required": [
          "id",
          "booking_id",
          "checked_out_by_user_id",
          "checked_out_to_user_id",
          "status",
          "notes"
        ],
        "title": "CheckoutSummaryRead",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": ["loc", "msg", "type"],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Inventory Booking API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/bookings": {
      "get": {
        "operationId": "list_booking_endpoint_bookings_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/BookingSummaryRead"
                  },
                  "title": "Response List Booking Endpoint Bookings Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Booking Endpoint",
        "tags": ["bookings"]
      },
      "post": {
        "operationId": "create_booking_endpoint_bookings_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookingCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Booking Endpoint",
        "tags": ["bookings"]
      }
    },
    "/bookings/availability": {
      "post": {
        "operationId": "preview_availability_endpoint_bookings_availability_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookingCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AvailabilityRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Preview Availability Endpoint",
        "tags": ["bookings"]
      }
    },
    "/bookings/availability/days": {
      "get": {
        "operationId": "availability_days_endpoint_bookings_availability_days_get",
        "parameters": [
          {
            "in": "query",
            "name": "asset_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Asset Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "starts_at",
            "required": true,
            "schema": {
              "format": "date-time",
              "title": "Starts At",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "ends_at",
            "required": true,
            "schema": {
              "format": "date-time",
              "title": "Ends At",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "quantity",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Quantity",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "location_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Location Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AvailabilityDaysRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Availability Days Endpoint",
        "tags": ["bookings"]
      }
    },
    "/bookings/availability/heatmap": {
      "get": {
        "operationId": "availability_heatmap_endpoint_bookings_availability_heatmap_get",
        "parameters": [
          {
            "in": "query",
            "name": "starts_at",
            "required": true,
            "schema": {
              "format": "date-time",
              "title": "Starts At",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "ends_at",
            "required": true,
            "schema": {
              "format": "date-time",
              "title": "Ends At",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "bucket",
            "required": false,
            "schema": {
              "default": "day",
              "pattern": "^(day|week)$",
              "title": "Bucket",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "location_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Location Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AvailabilityHeatmapRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Availability Heatmap Endpoint",
        "tags": ["bookings"]
      }
    },
    "/bookings/{booking_id}": {
      "delete": {
        "operationId": "delete_booking_endpoint_bookings__booking_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "booking_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Booking Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Booking Endpoint",
        "tags": ["bookings"]
      },
      "get": {
        "operationId": "get_booking_endpoint_bookings__booking_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "booking_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Booking Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Booking Endpoint",
        "tags": ["bookings"]
      },
      "patch": {
        "operationId": "update_booking_endpoint_bookings__booking_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "booking_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Booking Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookingUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Booking Endpoint",
        "tags": ["bookings"]
      }
    },
    "/bookings/{booking_id}/cancel": {
      "post": {
        "operationId": "cancel_booking_endpoint_bookings__booking_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "booking_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Booking Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel Booking Endpoint",
        "tags": ["bookings"]
      }
    },
...
    "/checkouts": {
      "get": {
        "operationId": "list_checkout_endpoint_checkouts_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CheckoutSummaryRead"
                  },
                  "title": "Response List Checkout Endpoint Checkouts Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Checkout Endpoint",
        "tags": ["checkouts"]
      },
      "post": {
        "operationId": "create_checkout_endpoint_checkouts_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckoutCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckoutRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Checkout Endpoint",
        "tags": ["checkouts"]
      }
    },
    "/checkouts/{checkout_id}": {
      "get": {
        "operationId": "get_checkout_endpoint_checkouts__checkout_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "checkout_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Checkout Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckoutRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Checkout Endpoint",
        "tags": ["checkouts"]
      }
    },
...
  }
}